+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - `album`:
   - Clean up/remove artist such as **Various**.
+  - Keep the artist or label in place when it's followed by part name, like **Volume**
 
 - `catalognum`:
   - Prevent matches followed by lowercase **ep**, **lp** and **va**.